+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.14.0] - 2022-02-23
+
+- Adds an index on device_id_hash to the codes table.
+- Using lower transaction isolation level while creating passwordless device with code
+
 ## [1.13.2] - 2022-02-19
 
 - Refactor Query Mechanism to avoid Memory Leaks